refactor(cli): split launch/mod.rs into util submodules
