+++ crates/cli/src/commands/util/mod.rs
@@ -1,0 +1,3 @@
+pub mod session_log;
+
+pub use session_log::*;
