+++ crates/cli/src/commands/util/session_log/mod.rs
@@ -1,0 +1,5 @@
+mod display;
+mod log;
+
+pub use display::*;
+pub use log::*;
